server/ocr-service/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "3.13"
anyio==4.6.2.post1 ; python_version >= "3.12" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.12" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.13" and platform_system == "Windows"
coloredlogs==15.0.1 ; python_version >= "3.12" and python_version < "3.13"
fastapi==0.115.6 ; python_version >= "3.12" and python_version < "3.13"
flatbuffers==24.3.25 ; python_version >= "3.12" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.12" and python_version < "3.13"
humanfriendly==10.0 ; python_version >= "3.12" and python_version < "3.13"
idna==3.10 ; python_version >= "3.12" and python_version < "3.13"
mpmath==1.3.0 ; python_version >= "3.12" and python_version < "3.13"
numpy==2.1.3 ; python_version >= "3.12" and python_version < "3.13"
ocrmac==1.0.0 ; python_version >= "3.12" and python_version < "3.13" and sys_platform == "darwin"
onnxruntime==1.20.1 ; python_version >= "3.12" and python_version < "3.13"
opencv-python==4.10.0.84 ; python_version >= "3.12" and python_version < "3.13"
packaging==24.2 ; python_version >= "3.12" and python_version < "3.13"
pillow==11.0.0 ; python_version >= "3.12" and python_version < "3.13"
protobuf==5.29.0 ; python_version >= "3.12" and python_version < "3.13"
py-cpuinfo==9.0.0 ; python_version >= "3.12" and python_version < "3.13"
pyclipper==1.3.0.post6 ; python_version >= "3.12" and python_version < "3.13"
pydantic-core==2.27.1 ; python_version >= "3.12" and python_version < "3.13"
pydantic-settings==2.6.1 ; python_version >= "3.12" and python_version < "3.13"
pydantic==2.10.3 ; python_version >= "3.12" and python_version < "3.13"
pyobjc-core==10.3.2 ; python_version >= "3.12" and python_version < "3.13" and sys_platform == "darwin"
pyobjc-framework-cocoa==10.3.2 ; python_version >= "3.12" and python_version < "3.13" and sys_platform == "darwin"
pyobjc-framework-coreml==10.3.2 ; python_version >= "3.12" and python_version < "3.13" and sys_platform == "darwin"
pyobjc-framework-quartz==10.3.2 ; python_version >= "3.12" and python_version < "3.13" and sys_platform == "darwin"
pyobjc-framework-vision==10.3.2 ; python_version >= "3.12" and python_version < "3.13" and sys_platform == "darwin"
pyreadline3==3.5.4 ; sys_platform == "win32" and python_version >= "3.12" and python_version < "3.13"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "3.13"
python-multipart==0.0.19 ; python_version >= "3.12" and python_version < "3.13"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "3.13"
rapidocr-onnxruntime==1.4.1 ; python_version >= "3.12" and python_version < "3.13"
rapidocr-paddle==1.4.1 ; python_version >= "3.12" and python_version < "3.13"
shapely==2.0.6 ; python_version >= "3.12" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.12" and python_version < "3.13"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "3.13"
starlette==0.41.3 ; python_version >= "3.12" and python_version < "3.13"
sympy==1.13.3 ; python_version >= "3.12" and python_version < "3.13"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "3.13"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "3.13"
uvicorn==0.32.1 ; python_version >= "3.12" and python_version < "3.13"
